requirements.txt
streamlit == 0.84.2
pandas == 1.1.4
numpy == 1.18.5
plotly.express == 5.1.0
pydeck == 0.6.2
